Enable checks for existing record before removing
